refactor(BoardInitializer) : 보드 초기화 추상화
