+++ src/main/java/chess/domain/ChessGame.java
@@ -1,5 +1,6 @@
 package chess.domain;
 
+import chess.domain.board.Board;
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
 import java.util.ArrayList;
