+++ src/main/java/chess/domain/board/BoardInitializer.java
@@ -1,0 +1,5 @@
+package chess.domain.board;
+
+public interface BoardInitializer {
+    Board initialize();
+}
