+++ src/main/java/chess/domain/board/MysqlBoardInitializer.java
@@ -1,9 +1,9 @@
-package chess.domain.dbUtils;
+package chess.domain.board;
 
-import chess.domain.Board;
 import chess.domain.Color;
 import chess.domain.Piece;
 import chess.domain.PieceType;
+import chess.domain.dbUtils.DBConnectionUtils;
 import chess.domain.position.Column;
 import chess.domain.position.Position;
 import chess.domain.position.Row;
@@ -14,14 +14,19 @@ import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.NoSuchElementException;
 
-public class MysqlBoardInitializer { // TODO 인터페이스 분리
+public class MysqlBoardInitializer implements BoardInitializer{
 
     private final Connection connection;
+
+    public MysqlBoardInitializer() {
+        this.connection = DBConnectionUtils.getConnection();
+    }
 
     public MysqlBoardInitializer(Connection connection) {
         this.connection = connection;
     }
 
+    @Override
     public Board initialize() {
         Board board = new Board(new HashMap<>());
         try {
@@ -40,20 +45,20 @@ public class MysqlBoardInitializer { // TODO 인터페이스 분리
         }
         return board;
     }
-
-    public void initPiece(Position position, Piece piece) {
-        final var query = "INSERT INTO board VALUES(?, ?, ?, ?)";
-        try {
-            final var preparedStatement = connection.prepareStatement(query);
-            preparedStatement.setString(1, piece.getPieceType().name());
-            preparedStatement.setInt(2, position.getColumnIndex());
-            preparedStatement.setInt(3, position.getRowIndex());
-            preparedStatement.setString(4, piece.getColor().name());
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
+//
+//    public void initPiece(Position position, Piece piece) {
+//        final var query = "INSERT INTO board VALUES(?, ?, ?, ?)";
+//        try {
+//            final var preparedStatement = connection.prepareStatement(query);
+//            preparedStatement.setString(1, piece.getPieceType().name());
+//            preparedStatement.setInt(2, position.getColumnIndex());
+//            preparedStatement.setInt(3, position.getRowIndex());
+//            preparedStatement.setString(4, piece.getColor().name());
+//            preparedStatement.executeUpdate();
+//        } catch (SQLException e) {
+//            throw new RuntimeException(e);
+//        }
+//    }
 
 
 //    public void addPiece(Position position, Piece piece) {
