+++ src/test/java/chess/domain/chessGame/BishopTest.java
@@ -2,7 +2,7 @@ package chess.domain.chessGame;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.domain.ChessGame;
 import chess.domain.Color;
 import chess.domain.position.Column;
